pyproject.toml
[project]
name = "llm-litsearch"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "instructor>=1.8.1",
    "openai>=1.78.1",
    "polars>=1.29.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "mypy>=1.15.0",
    "ruff>=0.11.9",
]
